+++ CHANGELOG.md
@@ -4,15 +4,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- Fix analyzers for Name and Full Name for search
 - Add -H/--homepage flag to open command
 - Add spinner for color output
 - Add language to the full-text search
 - Add tags to the full-text search
 
+### Fixed
+- Fix analyzers for Name and Full Name for search
+
 ## [0.2.0]
 ### Added
 - Full-text search with the search command
+
+### Fixed
 - Fix: List by tag no longer creates tag
 
 ## 0.1.0 - 2016-06-27