+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.3.0]
 ### Added
 - Add -H/--homepage flag to open command
 - Add spinner for color output
@@ -22,5 +24,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2016-06-27
 ### Initial Release
 
-[Unreleased]: https://github.com/hoop33/limo/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/hoop33/limo/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/hoop33/limo/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/hoop33/limo/compare/v0.1.0...v0.2.0